python/api-client/requirements.txt
python-dateutil
six
urllib3
vcd-api-schemas-type
PyYAML
